Multiple Choice widget and styling
Added a multiple choice widget for users to select scale 1-7 (not dynamic) I thought this would  be easier than making it dynamic but it wasnt and was a big mistake but its there now so....

The other thing I added was styling to the answers history, making answers a little bigger, not sure if I like it though. Might need a divider between the items.